+++ frontend/api.py
@@ -76,6 +76,7 @@ API_ALLOWED_METHODS = {
     'set_audio_muted',
     'get_cab_mode',
     'get_theme_assets_port',
+    'get_managerui_remote_link',
     'get_theme_index_page',
     'send_event',
     'send_event_all_windows',
@@ -404,6 +405,9 @@ class API:
 
     def get_theme_assets_port(self):
         return config_api.get_theme_assets_port(self._iniConfig.config)
+
+    def get_managerui_remote_link(self):
+        return config_api.get_managerui_remote_link(self._iniConfig.config)
 
     def get_theme_index_page(self):
         return theme_api.get_theme_index_page(self._iniConfig.config, self.get_my_window_name())