+++ RELEASE_NOTES.md
@@ -4,7 +4,8 @@
 This is the offical release of vpinfe v2.0.3
 
 ### What's New
-- More managerUI clearing.  Logs now on its own page.  added vpx logs
+- More managerUI clearing/seperation.  Logs now on its own page.  added vpx logs
+- QRCode in the mainmenu for linking to remote.
 
 ### Fixes
 None